backend/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --format requirements-txt --no-hashes
annotated-types==0.7.0
    # via
    #   float-backend
    #   pydantic
anyio==4.11.0
    # via
    #   httpx
    #   openai
boto3==1.40.64
    # via float-backend
botocore==1.40.64
    # via
    #   boto3
    #   s3transfer
cachetools==5.4.0
    # via
    #   float-backend
    #   google-auth
certifi==2024.7.4
    # via
    #   float-backend
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via
    #   float-backend
    #   requests
colorama==0.4.6 ; sys_platform == 'win32'
    # via tqdm
distro==1.9.0
    # via openai
google-ai-generativelanguage==0.6.6
    # via
    #   float-backend
    #   google-generativeai
google-api-core==2.19.1
    # via
    #   float-backend
    #   google-ai-generativelanguage
    #   google-api-python-client
    #   google-cloud-texttospeech
    #   google-generativeai
google-api-python-client==2.141.0
    # via
    #   float-backend
    #   google-generativeai
google-auth==2.33.0
    # via
    #   float-backend
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-cloud-texttospeech
    #   google-generativeai
google-auth-httplib2==0.2.0
    # via
    #   float-backend
    #   google-api-python-client
google-cloud-texttospeech==2.16.5
    # via float-backend
google-generativeai==0.7.2
    # via float-backend
googleapis-common-protos==1.63.2
    # via
    #   float-backend
    #   google-api-core
    #   grpcio-status
grpcio==1.65.4
    # via
    #   float-backend
    #   google-api-core
    #   grpcio-status
grpcio-status==1.62.3
    # via
    #   float-backend
    #   google-api-core
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httplib2==0.22.0
    # via
    #   float-backend
    #   google-api-python-client
    #   google-auth-httplib2
httpx==0.28.1
    # via openai
idna==3.7
    # via
    #   anyio
    #   float-backend
    #   httpx
    #   requests
jinja2==3.1.6
    # via float-backend
jiter==0.11.1
    # via openai
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
markupsafe==3.0.3
    # via jinja2
openai==2.6.1
    # via float-backend
proto-plus==1.24.0
    # via
    #   float-backend
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-cloud-texttospeech
protobuf==4.25.4
    # via
    #   float-backend
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-cloud-texttospeech
    #   google-generativeai
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.0
    # via
    #   float-backend
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.0
    # via
    #   float-backend
    #   google-auth
pydantic==2.8.2
    # via
    #   float-backend
    #   google-generativeai
    #   openai
pydantic-core==2.20.1
    # via
    #   float-backend
    #   pydantic
pydub==0.25.1
    # via float-backend
pyparsing==3.1.2
    # via
    #   float-backend
    #   httplib2
python-dateutil==2.9.0.post0
    # via botocore
python-dotenv==1.0.1
    # via float-backend
requests==2.32.5
    # via
    #   float-backend
    #   google-api-core
rsa==4.9
    # via
    #   float-backend
    #   google-auth
s3transfer==0.14.0
    # via boto3
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   openai
tqdm==4.66.5
    # via
    #   float-backend
    #   google-generativeai
    #   openai
typing-extensions==4.12.2
    # via
    #   anyio
    #   float-backend
    #   google-generativeai
    #   openai
    #   pydantic
    #   pydantic-core
uritemplate==4.1.1
    # via
    #   float-backend
    #   google-api-python-client
urllib3==2.5.0
    # via
    #   botocore
    #   float-backend
    #   requests
